@@ -15,120 +15,119 @@
   "dependencies": {
     "apollo-client": "^1.0.0-rc.2",
     "aws-sdk": "^2.38.0",
+    "babel-cli": "^6.24.1",
+    "babel-core": "^6.24.1",
+    "babel-loader": "^7.1.2",
+    "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-plugin-transform-runtime": "^6.23.0",
     "babel-polyfill": "^6.23.0",
     "babel-preset-es2015": "^6.18.0",
+    "babel-preset-react": "^6.24.1",
     "babel-preset-stage-3": "^6.17.0",
     "babel-runtime": "^6.20.0",
     "bcryptjs": "^2.4.0",
-    "body-parser": "~1.15.2",
+    "body-parser": "~1.18.2",
     "camelcase": "^4.0.0",
+    "compression-webpack-plugin": "^1.0.1",
     "connect-redis": "^3.2.0",
     "cookie-parser": "~1.4.3",
-    "debug": "^2.6.0",
+    "css-loader": "^0.28.7",
+    "csv-parse": "^1.2.0",
+    "debug": "^3.1.0",
     "decamelize": "^1.2.0",
+    "del": "~3.0.0",
     "didyoumean": "^1.2.1",
     "ejs": "^2.5.5",
-    "express": "~4.14.0",
+    "express": "~4.16.2",
     "express-graphql": "^0.6.2",
     "express-session": "^1.14.2",
-    "graphql": "^0.9.1",
-    "graphql-server-express": "^0.6.0",
-    "graphql-tag": "^2.1.0",
-    "gulp-sass": "~3.0.0",
-    "history": "^4.5.1",
-    "ical-generator": "^0.2.9",
-    "knex": "^0.12.6",
-    "lodash.isequal": "^4.5.0",
-    "lodash.uniqwith": "^4.5.0",
-    "moment": "2.18.1",
-    "moment-timezone": "^0.5.13",
-    "morgan": "~1.7.0",
-    "node-sass": "^4.5.3",
-    "nodemon": "^1.11.0",
-    "normalize.css": "^5.0.0",
-    "passport": "^0.3.2",
-    "passport-local": "^1.0.0",
-    "pg": "^6.1.2",
-    "pluralize": "^4.0.0",
-    "postmark": "^1.3.1",
-    "query-string": "^4.3.0",
-    "raven": "^2.0.2",
-    "raven-js": "^3.15.0",
-    "react-apollo": "^1.0.0-rc.2",
-    "redis": "^2.6.5",
-    "rss": "^1.2.2",
-    "speakingurl": "^13.0.0",
-    "uuid": "^3.0.1",
-    "validator": "^6.2.1",
-    "webpack-stream": "^4.0.0",
-    "zipcode-to-timezone": "0.0.8"
-  },
-  "devDependencies": {
-    "babel-cli": "^6.24.1",
-    "babel-core": "^6.24.1",
-    "babel-eslint": "^7.2.3",
-    "babel-loader": "^6.4.1",
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-preset-react": "^6.24.1",
-    "bootstrap-sass": "^3.3.7",
-    "chai": "^3.5.0",
-    "chai-enzyme": "^0.6.1",
-    "compression-webpack-plugin": "^0.3.2",
-    "css-loader": "^0.26.4",
-    "csv-parse": "^1.2.0",
-    "del": "~2.2.2",
-    "enzyme": "^2.8.2",
-    "eslint": "^4.2.0",
-    "eslint-config-airbnb": "^15.0.2",
-    "eslint-loader": "^1.9.0",
-    "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-jsx-a11y": "^6.0.2",
-    "eslint-plugin-react": "^7.1.0",
-    "extract-text-webpack-plugin": "^2.1.0",
+    "extract-text-webpack-plugin": "^3.0.1",
     "favicons-webpack-plugin": "0.0.7",
-    "file-loader": "^0.9.0",
+    "file-loader": "^1.1.5",
     "front-matter": "^2.1.2",
+    "graphql": "^0.11.7",
+    "graphql-server-express": "^1.1.6",
+    "graphql-tag": "^2.1.0",
     "gulp": "^3.9.1",
     "gulp-clean": "~0.3.2",
     "gulp-environments": "^0.1.2",
     "gulp-livereload": "^3.8.1",
     "gulp-nodemon": "^2.2.1",
-    "gulp-notify": "~2.2.0",
+    "gulp-notify": "~3.0.0",
     "gulp-plumber": "^1.1.0",
+    "gulp-sass": "~3.1.0",
     "gulp-watch": "^4.3.11",
+    "history": "^4.5.1",
     "html-webpack-plugin": "^2.28.0",
     "html-webpack-template": "^6.0.1",
+    "ical-generator": "^0.2.9",
+    "knex": "^0.13.0",
+    "lodash.isequal": "^4.5.0",
+    "lodash.uniqwith": "^4.5.0",
+    "moment": "2.19.0",
+    "moment-timezone": "^0.5.13",
+    "morgan": "~1.9.0",
+    "node-sass": "^4.5.3",
+    "nodemon": "^1.11.0",
+    "normalize.css": "^7.0.0",
+    "papaparse": "^4.3.3",
+    "passport": "^0.4.0",
+    "passport-local": "^1.0.0",
+    "pg": "^7.3.0",
+    "pluralize": "^7.0.0",
+    "postmark": "^1.3.1",
+    "prop-types": "^15.6.0",
+    "query-string": "^5.0.0",
+    "raven": "^2.0.2",
+    "raven-js": "^3.15.0",
+    "react": "^16.0.0",
+    "react-apollo": "^1.0.0-rc.2",
+    "react-dom": "^16.0.0",
+    "react-dropzone": "^4.2.0",
+    "react-ga": "^2.2.0",
+    "react-image-crop": "^3.0.2",
+    "react-infinite": "^0.12.1",
+    "react-redux": "^5.0.4",
+    "react-tap-event-plugin": "^3.0.2",
+    "react-virtualized": "^9.9.0",
+    "redis": "^2.6.5",
+    "redux": "^3.6.0",
+    "redux-logger": "^3.0.6",
+    "redux-thunk": "^2.2.0",
+    "resolve-url-loader": "^2.1.1",
+    "rss": "^1.2.2",
+    "sass-loader": "^6.0.6",
+    "speakingurl": "^14.0.1",
+    "style-loader": "^0.19.0",
+    "sw-precache-webpack-plugin": "git://github.com/max-b/sw-precache-webpack-plugin.git#prepare",
+    "universal-router": "^4.2.1",
+    "url-loader": "^0.6.2",
+    "uuid": "^3.0.1",
+    "validator": "^9.0.0",
+    "webpack": "^3.6.0",
+    "webpack-pwa-manifest": "^3.3.2",
+    "webpack-stream": "^4.0.0",
+    "zipcode-to-timezone": "0.0.9"
+  },
+  "devDependencies": {
+    "babel-eslint": "^8.0.1",
+    "chai": "^4.1.2",
+    "chai-enzyme": "^0.8.0",
+    "enzyme": "^3.0.0",
+    "eslint": "^4.8.0",
+    "eslint-config-airbnb": "^16.0.0",
+    "eslint-loader": "^1.9.0",
+    "eslint-plugin-import": "^2.7.0",
+    "eslint-plugin-jsx-a11y": "^6.0.2",
+    "eslint-plugin-react": "^7.4.0",
     "ignore-styles": "^5.0.1",
     "isomorphic-fetch": "^2.2.1",
-    "jsdom": "^9.12.0",
+    "jsdom": "^11.3.0",
     "markdown-it": "^8.3.1",
-    "material-ui": "^0.18.4",
-    "mocha": "^3.3.0",
-    "papaparse": "^4.3.3",
-    "react": "^15.5.4",
+    "material-ui": "^0.19.4",
+    "mocha": "^4.0.1",
     "react-addons-test-utils": "^15.5.1",
-    "react-bootstrap": "^0.30.10",
-    "react-dom": "^15.5.4",
-    "react-dropzone": "^3.13.1",
-    "react-ga": "^2.2.0",
-    "react-image-crop": "^2.0.3",
-    "react-infinite": "^0.11.0",
-    "react-redux": "^5.0.4",
-    "react-tap-event-plugin": "^2.0.1",
-    "react-virtualized": "^9.9.0",
     "redbox-react": "^1.3.6",
-    "redux": "^3.6.0",
-    "redux-logger": "^2.10.2",
-    "redux-thunk": "^2.2.0",
-    "resolve-url-loader": "^1.6.1",
-    "sass-loader": "^4.1.1",
-    "style-loader": "^0.13.2",
-    "sw-precache-webpack-plugin": "git://github.com/max-b/sw-precache-webpack-plugin.git#prepare",
-    "universal-router": "^2.0.0",
-    "url-loader": "^0.5.8",
-    "webpack": "^2.4.1",
-    "webpack-bundle-analyzer": "^2.4.0",
-    "webpack-pwa-manifest": "^3.3.2"
+    "webpack-bundle-analyzer": "^2.4.0"
   }
 }
