@@ -1,4 +1,5 @@
-import React, { PureComponent, PropTypes } from 'react';
+import PropTypes from 'prop-types';
+import React, { PureComponent } from 'react';
 import Infinite from 'react-infinite';
 import moment from 'moment';
 import isEqual from 'lodash.isequal';
