@@ -1,4 +1,5 @@
-import React, { PureComponent, PropTypes } from 'react';
+import PropTypes from 'prop-types';
+import React, { PureComponent } from 'react';
 import FontIcon from 'material-ui/FontIcon';
 
 import Link from 'components/Link';
